+++ src/sites/image/imagevenue.js
@@ -3,7 +3,7 @@ _.register({
     'https://*.imagevenue.com/*',
   ],
   async ready () {
-    const i = $('a[title] img');
+    const i = $('a[title] img#main-image');
     await $.openImage(i.src);
   },
 });