+++ src/sites/link/earnlink.io.js
@@ -1,0 +1,9 @@
+_.register({
+  rule: {
+    host: /^earnlink\.io$/,
+  },
+  async ready() {
+    const directUrl = $.searchFromScripts(/"([^"]+)"\)\.html\("Continue"\)/);
+    await $.openLink(directUrl[1]);
+  },
+});